refactor: Split pybigtools lib into submodules
